requirements.txt
fastapi
uvicorn
httpx
aiofiles
pyyaml
vllm
transformers